Added several new modals
A few updates to both pages. The general intent at the moment is to align different profile information into the single JSON file. Eventually, crud ability will be added to that JSON file (and eventually database), but the current iteration is spent figuring out what *will* connect and creating the components around that.

To that effect, here are the update notes.

Added new modals to Profile List:
- Add user: pops up unformatted modal to add new user based on email and role permissions. Still need to add more fields, organize and format this window.
- User info pop up: info about user pops up when selecting their row. Unfortunately, there is a bug that selecting the checkbox also selects this.

Slightly modified Profile page:
- Edit details button but before more work done here, I need to add alternates to textfield depending on whether profile is in edit state or not. This will be a little bit of a pain but nice once it is done!