Reorganized and made Header
Moved a lot of the details page files around to better organize. Also fixed up routing for that.

Also made Header component to act as head of all pages. Included button coloring for navbar. Still need to fix the profile/profiles page as well as fixing the rounding on the buttons.